Refactor telemetry package
