+++ backend/internal/pkg/telemetry/client.go
@@ -3,17 +3,9 @@ package telemetry
 import (
 	"context"
 	"errors"
-	"time"
-
-	"go.opentelemetry.io/otel/sdk/resource"
-
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/exporters/stdout/stdoutmetric"
-	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 )
 
 const otelName = "Plex Recommendation API"
@@ -22,25 +14,6 @@ var (
 	Tracer = otel.Tracer(otelName)
 	Meter  = otel.Meter(otelName)
 )
-
-type telemetryOptions struct {
-	withTracer bool
-	withMeter  bool
-}
-
-type Option func(*telemetryOptions)
-
-func WithTracer(enabled bool) Option {
-	return func(o *telemetryOptions) {
-		o.withTracer = enabled
-	}
-}
-
-func WithMeter(enabled bool) Option {
-	return func(o *telemetryOptions) {
-		o.withMeter = enabled
-	}
-}
 
 // InitOtel bootstraps the OpenTelemetry pipeline.
 // If it does not return an error, make sure to call shutdown for proper cleanup.
@@ -99,46 +72,4 @@ func InitOtel(ctx context.Context, opts ...Option) (func(context.Context) error,
 	}
 
 	return shutdown, nil
-}
-
-func newPropagator() propagation.TextMapPropagator {
-	return propagation.NewCompositeTextMapPropagator(
-		propagation.TraceContext{},
-		propagation.Baggage{},
-	)
-}
-
-func newTraceProvider() (*trace.TracerProvider, error) {
-
-	exporter, err := otlptracegrpc.New(context.Background(), otlptracegrpc.WithInsecure())
-	if err != nil {
-		return nil, err
-	}
-
-	traceProvider := trace.NewTracerProvider(
-		trace.WithBatcher(
-			exporter,
-			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
-			trace.WithBatchTimeout(trace.DefaultScheduleDelay*time.Millisecond),
-			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
-		),
-		trace.WithResource(
-			resource.NewWithAttributes(
-				semconv.SchemaURL,
-				semconv.ServiceNameKey.String("Plex Recommendation Backend"),
-			),
-		),
-	)
-	return traceProvider, nil
-}
-
-func newMeterProvider() (*metric.MeterProvider, error) {
-	metricExporter, err := stdoutmetric.New()
-	if err != nil {
-		return nil, err
-	}
-
-	meterProvider := metric.NewMeterProvider(
-		metric.WithReader(metric.NewPeriodicReader(metricExporter)))
-	return meterProvider, nil
 }
