+++ backend/internal/pkg/telemetry/options.go
@@ -1,0 +1,20 @@
+package telemetry
+
+type telemetryOptions struct {
+	withTracer bool
+	withMeter  bool
+}
+
+type Option func(*telemetryOptions)
+
+func WithTracer(enabled bool) Option {
+	return func(o *telemetryOptions) {
+		o.withTracer = enabled
+	}
+}
+
+func WithMeter(enabled bool) Option {
+	return func(o *telemetryOptions) {
+		o.withMeter = enabled
+	}
+}
